Implements worker based textmate tokenization. Adds setting "editor.experimental.asyncTokenization" (default false).
