+++ src/vs/workbench/contrib/codeEditor/electron-sandbox/startDebugTextMate.ts
@@ -7,7 +7,7 @@ import * as nls from 'vs/nls';
 import { Range } from 'vs/editor/common/core/range';
 import { Action2, registerAction2 } from 'vs/platform/actions/common/actions';
 import { Categories } from 'vs/platform/action/common/actionCommonCategories';
-import { ITextMateService } from 'vs/workbench/services/textMate/browser/textMate';
+import { ITextMateTokenizationFeature } from 'vs/workbench/services/textMate/browser/textMateTokenizationFeature';
 import { IModelService } from 'vs/editor/common/services/model';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
 import { URI } from 'vs/base/common/uri';
@@ -52,7 +52,7 @@ class StartDebugTextMate extends Action2 {
 	}
 
 	async run(accessor: ServicesAccessor) {
-		const textMateService = accessor.get(ITextMateService);
+		const textMateService = accessor.get(ITextMateTokenizationFeature);
 		const modelService = accessor.get(IModelService);
 		const editorService = accessor.get(IEditorService);
 		const codeEditorService = accessor.get(ICodeEditorService);
