+++ src/vs/workbench/services/textMate/common/TMScopeRegistry.ts
@@ -5,7 +5,6 @@
 
 import * as resources from 'vs/base/common/resources';
 import { URI } from 'vs/base/common/uri';
-import { Disposable } from 'vs/base/common/lifecycle';
 import { LanguageId, StandardTokenType } from 'vs/editor/common/encodedTokenAttributes';
 
 export interface IValidGrammarDefinition {
@@ -27,12 +26,11 @@ export interface IValidEmbeddedLanguagesMap {
 	[scopeName: string]: LanguageId;
 }
 
-export class TMScopeRegistry extends Disposable {
+export class TMScopeRegistry {
 
 	private _scopeNameToLanguageRegistration: { [scopeName: string]: IValidGrammarDefinition };
 
 	constructor() {
-		super();
 		this._scopeNameToLanguageRegistration = Object.create(null);
 	}
